Move subscription state into a separate service
